FE-1218 - move function to shared export
