+++ src/pages/domains/components/TrackingDomainsTab.js
@@ -10,6 +10,7 @@ import { API_ERROR_MESSAGE } from '../constants';
 import useDomains from '../hooks/useDomains';
 import TableFilters, { reducer as tableFiltersReducer } from './TableFilters';
 import TrackingDomainsTable from './TrackingDomainsTable';
+import getReactTableFilters from './getReactTableFilters';
 
 const filtersInitialState = {
   domainName: undefined,
@@ -52,36 +53,6 @@ const initFiltersForTracking = {
       return val === 'true' || val === 'false' || typeof val === 'boolean';
     },
   },
-};
-
-// TODO: Move and import here
-const getReactTableFilters = (
-  filters,
-  filtersState,
-  { domainName = null, targetName = undefined, targetChecked = undefined },
-) => {
-  const newFilters = Object.keys(filters).map(i => {
-    if (domainName && 'domainName' === i) {
-      return { id: 'domainName', value: domainName };
-    } else if (targetName && targetChecked !== undefined && targetName === i) {
-      return { id: i, value: targetChecked };
-    }
-    let index = filtersState.checkboxes.map(i => i.name).indexOf(i);
-    let filter = filtersState.checkboxes[index];
-
-    if (!filter) {
-      return null;
-    } else if (filter.isChecked) {
-      return {
-        id: i,
-        value: true,
-      };
-    }
-
-    return null;
-  });
-
-  return newFilters.filter(Boolean);
 };
 
 export default function TrackingDomainsTab() {
